+++ model-engine/tests/unit/conftest.py
@@ -14,6 +14,7 @@ from typing import (
     Set,
     Tuple,
 )
+from unittest import mock
 from unittest.mock import mock_open
 from uuid import uuid4
 
@@ -4535,3 +4536,57 @@ def llm_model_endpoint_trt_llm(
             image="test_image",
         ),
     )
+
+
+def mocked__get_recommended_hardware_config_map():
+    async def async_mock(*args, **kwargs):  # noqa
+        return {
+            "byGpuMemoryGb": """
+    - gpu_memory_le: 20
+      cpus: 5
+      gpus: 1
+      memory: 20Gi
+      storage: 40Gi
+      gpu_type: nvidia-hopper-h100-1g20gb
+    - gpu_memory_le: 40
+      cpus: 10
+      gpus: 1
+      memory: 40Gi
+      storage: 80Gi
+      gpu_type: nvidia-hopper-h100-3g40gb
+    - gpu_memory_le: 80
+      cpus: 20
+      gpus: 1
+      memory: 80Gi
+      storage: 96Gi
+      gpu_type: nvidia-hopper-h100
+    - gpu_memory_le: 160
+      cpus: 40
+      gpus: 2
+      memory: 160Gi
+      storage: 160Gi
+      gpu_type: nvidia-hopper-h100
+    - gpu_memory_le: 320
+      cpus: 80
+      gpus: 4
+      memory: 320Gi
+      storage: 320Gi
+      gpu_type: nvidia-hopper-h100
+    - gpu_memory_le: 640
+      cpus: 160
+      gpus: 8
+      memory: 800Gi
+      storage: 640Gi
+      gpu_type: nvidia-hopper-h100
+                """,
+            "byModelName": """
+    - name: llama-3-8b-instruct-262k
+      cpus: 40
+      gpus: 2
+      memory: 160Gi
+      storage: 160Gi
+      gpu_type: nvidia-hopper-h100
+                """,
+        }
+
+    return mock.AsyncMock(side_effect=async_mock)
